+++ example/convertToV3.ts
@@ -3,6 +3,8 @@
  * -d | --directory : map folder directory
  * -q | --quite : reduced log output (overridden by verbose)
  * -v | --verbose : enable debug log output
+ * -x | --no-backup : disable backup
+ * -y | --no-prompt : auto-complete prompt
  * example run command:
  * deno run --allow-read --allow-write convertToV3.ts -d "./Folder/Path"
  */
@@ -12,16 +14,21 @@ import { load, isV3, convert, save, logger, globals } from '../mod.ts';
 
 const args = parse(Deno.args, {
     string: 'd',
-    boolean: ['v', 'q'],
-    alias: { d: 'directory', q: 'quite', v: 'verbose' },
+    boolean: ['v', 'q', 'x', 'y'],
+    alias: { d: 'directory', q: 'quite', v: 'verbose', x: 'no-backup', y: 'no-prompt' },
 });
 
-logger.info('Beat Saber beatmap v2 to v3 conversion build 1');
+logger.info('Beat Saber beatmap v2 to v3 conversion build 2');
 logger.info('Source code available at https://github.com/KivalEvan/BeatSaber-Deno/blob/main/example/convertToV3.ts');
 logger.info('Send any feedback to Kival Evan#5480 on Discord');
 
+if (args.x) {
+    logger.warn('No backup flagged, any changes done by this script is irreversible');
+}
+
 globals.directory =
-    (args.d as string) ?? (prompt('Enter map folder path (leave blank for current folder):')?.trim() || './');
+    (args.d as string) ??
+    (args.y ? './' : prompt('Enter map folder path (leave blank for current folder):')?.trim() || './');
 
 if (args.q) {
     logger.setLevel(4);
@@ -53,31 +60,34 @@ try {
     let gradientChromaConfirm = false;
     diffList.forEach((dl) => {
         if (!isV3(dl.data)) {
-            logger.info('Backing up beatmap v2', dl.characteristic, dl.difficulty);
-            try {
-                copySync(
-                    globals.directory + dl.settings._beatmapFilename,
-                    globals.directory + dl.settings._beatmapFilename + '.old'
-                );
-            } catch (_) {
-                const confirmation = prompt('Old backup file detected, do you want to overwrite? (y/N):', 'n');
-                if (confirmation![0].toLowerCase() === 'y') {
+            if (!args.x) {
+                logger.info('Backing up beatmap v2', dl.characteristic, dl.difficulty);
+                try {
                     copySync(
                         globals.directory + dl.settings._beatmapFilename,
-                        globals.directory + dl.settings._beatmapFilename + '.old',
-                        { overwrite: true }
+                        globals.directory + dl.settings._beatmapFilename + '.old'
                     );
-                } else {
-                    logger.info('Skipping overwrite...');
-                    return;
+                } catch (_) {
+                    const confirmation = args.y
+                        ? 'n'
+                        : prompt('Old backup file detected, do you want to overwrite? (y/N):', 'n');
+                    if (confirmation![0].toLowerCase() === 'y') {
+                        copySync(
+                            globals.directory + dl.settings._beatmapFilename,
+                            globals.directory + dl.settings._beatmapFilename + '.old',
+                            { overwrite: true }
+                        );
+                    } else {
+                        logger.info('Skipping overwrite...');
+                        return;
+                    }
                 }
             }
             if (dl.data.events.some((e) => e.hasOldChroma())) {
                 if (!oldChromaConfirm) {
-                    const confirmation = prompt(
-                        'Old Chroma detected, do you want to convert this (apply to all)? (y/N):',
-                        'n'
-                    );
+                    const confirmation = args.y
+                        ? 'n'
+                        : prompt('Old Chroma detected, do you want to convert this (apply to all)? (y/N):', 'n');
                     if (confirmation![0].toLowerCase() === 'y') {
                         oldChromaConvert = true;
                     }
@@ -89,10 +99,12 @@ try {
             }
             if (dl.data.events.some((e) => e.customData._lightGradient)) {
                 if (!gradientChromaConfirm) {
-                    const confirmation = prompt(
-                        'Chroma light gradient detected, do you want to convert this (apply to all)? (y/N):',
-                        'n'
-                    );
+                    const confirmation = args.y
+                        ? 'n'
+                        : prompt(
+                              'Chroma light gradient detected, do you want to convert this (apply to all)? (y/N):',
+                              'n'
+                          );
                     if (confirmation![0].toLowerCase() === 'y') {
                         gradientChromaConvert = true;
                     }
@@ -116,11 +128,11 @@ try {
     }
 
     if (!args.d) {
-        prompt('Enter any key to exit...');
+        args.y || prompt('Enter any key to exit...');
     }
 } catch (e) {
     logger.error(e.message);
     logger.error('If this is an unexpected or unknown error');
     logger.error('Please report this to Kival Evan#5480 on Discord');
-    prompt('!! Enter any key to exit...');
+    args.y || prompt('!! Enter any key to exit...');
 }
